refactor: deprecate RepositoryProxy
