+++ composer.json
@@ -53,7 +53,7 @@
         "psr-4": {"Zenstruck\\Foundry\\": "src/"},
         "files": [
             "src/functions.php",
-            "src/deprecated.php",
+            "src/deprecations.php",
             "src/Persistence/functions.php"
         ]
     },
