+++ src/Bundle/Maker/Factory/NoPersistenceObjectsAutoCompleter.php
@@ -34,7 +34,7 @@ final class NoPersistenceObjectsAutoCompleter
 
                 $class = $this->toPSR4($rootPath, $phpFile, $namespacePrefix);
 
-                if ($class === 'Zenstruck\Foundry\Proxy') {
+                if (in_array($class, ['Zenstruck\Foundry\Proxy', 'Zenstruck\Foundry\RepositoryProxy'])) {
                     // do not load legacy Proxy: prevents deprecations in tests.
                     continue;
                 }
