--- src/deprecated.php
@@ -1,3 +1,0 @@
-<?php
-
-class_alias(\Zenstruck\Foundry\Persistence\Proxy::class, \Zenstruck\Foundry\Proxy::class);
