+++ tests/Fixtures/Maker/expected/can_create_factory_for_odm_with_data_set_document.php
@@ -14,27 +14,27 @@ namespace App\Factory;
 use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\Document\ODMPost;
 
 /**
  * @extends PersistentProxyObjectFactory<ODMPost>
  *
- * @method        ODMPost|Proxy                      create(array|callable $attributes = [])
- * @method static ODMPost|Proxy                      createOne(array $attributes = [])
- * @method static ODMPost|Proxy                      find(object|array|mixed $criteria)
- * @method static ODMPost|Proxy                      findOrCreate(array $attributes)
- * @method static ODMPost|Proxy                      first(string $sortedField = 'id')
- * @method static ODMPost|Proxy                      last(string $sortedField = 'id')
- * @method static ODMPost|Proxy                      random(array $attributes = [])
- * @method static ODMPost|Proxy                      randomOrCreate(array $attributes = [])
- * @method static DocumentRepository|RepositoryProxy repository()
- * @method static ODMPost[]|Proxy[]                  all()
- * @method static ODMPost[]|Proxy[]                  createMany(int $number, array|callable $attributes = [])
- * @method static ODMPost[]|Proxy[]                  createSequence(iterable|callable $sequence)
- * @method static ODMPost[]|Proxy[]                  findBy(array $attributes)
- * @method static ODMPost[]|Proxy[]                  randomRange(int $min, int $max, array $attributes = [])
- * @method static ODMPost[]|Proxy[]                  randomSet(int $number, array $attributes = [])
+ * @method        ODMPost|Proxy                          create(array|callable $attributes = [])
+ * @method static ODMPost|Proxy                          createOne(array $attributes = [])
+ * @method static ODMPost|Proxy                          find(object|array|mixed $criteria)
+ * @method static ODMPost|Proxy                          findOrCreate(array $attributes)
+ * @method static ODMPost|Proxy                          first(string $sortedField = 'id')
+ * @method static ODMPost|Proxy                          last(string $sortedField = 'id')
+ * @method static ODMPost|Proxy                          random(array $attributes = [])
+ * @method static ODMPost|Proxy                          randomOrCreate(array $attributes = [])
+ * @method static DocumentRepository|RepositoryDecorator repository()
+ * @method static ODMPost[]|Proxy[]                      all()
+ * @method static ODMPost[]|Proxy[]                      createMany(int $number, array|callable $attributes = [])
+ * @method static ODMPost[]|Proxy[]                      createSequence(iterable|callable $sequence)
+ * @method static ODMPost[]|Proxy[]                      findBy(array $attributes)
+ * @method static ODMPost[]|Proxy[]                      randomRange(int $min, int $max, array $attributes = [])
+ * @method static ODMPost[]|Proxy[]                      randomSet(int $number, array $attributes = [])
  */
 final class ODMPostFactory extends PersistentProxyObjectFactory
 {
