+++ tests/Fixtures/Maker/expected/can_create_factory_in_test_dir_interactively.php
@@ -14,27 +14,27 @@ namespace App\Tests\Factory;
 use Doctrine\ORM\EntityRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Tag;
 
 /**
  * @extends PersistentProxyObjectFactory<Tag>
  *
- * @method        Tag|Proxy                        create(array|callable $attributes = [])
- * @method static Tag|Proxy                        createOne(array $attributes = [])
- * @method static Tag|Proxy                        find(object|array|mixed $criteria)
- * @method static Tag|Proxy                        findOrCreate(array $attributes)
- * @method static Tag|Proxy                        first(string $sortedField = 'id')
- * @method static Tag|Proxy                        last(string $sortedField = 'id')
- * @method static Tag|Proxy                        random(array $attributes = [])
- * @method static Tag|Proxy                        randomOrCreate(array $attributes = [])
- * @method static EntityRepository|RepositoryProxy repository()
- * @method static Tag[]|Proxy[]                    all()
- * @method static Tag[]|Proxy[]                    createMany(int $number, array|callable $attributes = [])
- * @method static Tag[]|Proxy[]                    createSequence(iterable|callable $sequence)
- * @method static Tag[]|Proxy[]                    findBy(array $attributes)
- * @method static Tag[]|Proxy[]                    randomRange(int $min, int $max, array $attributes = [])
- * @method static Tag[]|Proxy[]                    randomSet(int $number, array $attributes = [])
+ * @method        Tag|Proxy                            create(array|callable $attributes = [])
+ * @method static Tag|Proxy                            createOne(array $attributes = [])
+ * @method static Tag|Proxy                            find(object|array|mixed $criteria)
+ * @method static Tag|Proxy                            findOrCreate(array $attributes)
+ * @method static Tag|Proxy                            first(string $sortedField = 'id')
+ * @method static Tag|Proxy                            last(string $sortedField = 'id')
+ * @method static Tag|Proxy                            random(array $attributes = [])
+ * @method static Tag|Proxy                            randomOrCreate(array $attributes = [])
+ * @method static EntityRepository|RepositoryDecorator repository()
+ * @method static Tag[]|Proxy[]                        all()
+ * @method static Tag[]|Proxy[]                        createMany(int $number, array|callable $attributes = [])
+ * @method static Tag[]|Proxy[]                        createSequence(iterable|callable $sequence)
+ * @method static Tag[]|Proxy[]                        findBy(array $attributes)
+ * @method static Tag[]|Proxy[]                        randomRange(int $min, int $max, array $attributes = [])
+ * @method static Tag[]|Proxy[]                        randomSet(int $number, array $attributes = [])
  */
 final class TagFactory extends PersistentProxyObjectFactory
 {
