+++ tests/Fixtures/Maker/expected/can_create_factory_with_default_enum_with_data_set_odm.php
@@ -14,28 +14,28 @@ namespace App\Factory;
 use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\DocumentWithEnum;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\SomeEnum;
 
 /**
  * @extends PersistentProxyObjectFactory<DocumentWithEnum>
  *
- * @method        DocumentWithEnum|Proxy             create(array|callable $attributes = [])
- * @method static DocumentWithEnum|Proxy             createOne(array $attributes = [])
- * @method static DocumentWithEnum|Proxy             find(object|array|mixed $criteria)
- * @method static DocumentWithEnum|Proxy             findOrCreate(array $attributes)
- * @method static DocumentWithEnum|Proxy             first(string $sortedField = 'id')
- * @method static DocumentWithEnum|Proxy             last(string $sortedField = 'id')
- * @method static DocumentWithEnum|Proxy             random(array $attributes = [])
- * @method static DocumentWithEnum|Proxy             randomOrCreate(array $attributes = [])
- * @method static DocumentRepository|RepositoryProxy repository()
- * @method static DocumentWithEnum[]|Proxy[]         all()
- * @method static DocumentWithEnum[]|Proxy[]         createMany(int $number, array|callable $attributes = [])
- * @method static DocumentWithEnum[]|Proxy[]         createSequence(iterable|callable $sequence)
- * @method static DocumentWithEnum[]|Proxy[]         findBy(array $attributes)
- * @method static DocumentWithEnum[]|Proxy[]         randomRange(int $min, int $max, array $attributes = [])
- * @method static DocumentWithEnum[]|Proxy[]         randomSet(int $number, array $attributes = [])
+ * @method        DocumentWithEnum|Proxy                 create(array|callable $attributes = [])
+ * @method static DocumentWithEnum|Proxy                 createOne(array $attributes = [])
+ * @method static DocumentWithEnum|Proxy                 find(object|array|mixed $criteria)
+ * @method static DocumentWithEnum|Proxy                 findOrCreate(array $attributes)
+ * @method static DocumentWithEnum|Proxy                 first(string $sortedField = 'id')
+ * @method static DocumentWithEnum|Proxy                 last(string $sortedField = 'id')
+ * @method static DocumentWithEnum|Proxy                 random(array $attributes = [])
+ * @method static DocumentWithEnum|Proxy                 randomOrCreate(array $attributes = [])
+ * @method static DocumentRepository|RepositoryDecorator repository()
+ * @method static DocumentWithEnum[]|Proxy[]             all()
+ * @method static DocumentWithEnum[]|Proxy[]             createMany(int $number, array|callable $attributes = [])
+ * @method static DocumentWithEnum[]|Proxy[]             createSequence(iterable|callable $sequence)
+ * @method static DocumentWithEnum[]|Proxy[]             findBy(array $attributes)
+ * @method static DocumentWithEnum[]|Proxy[]             randomRange(int $min, int $max, array $attributes = [])
+ * @method static DocumentWithEnum[]|Proxy[]             randomSet(int $number, array $attributes = [])
  */
 final class DocumentWithEnumFactory extends PersistentProxyObjectFactory
 {
