+++ tests/Fixtures/Maker/expected/can_create_factory_with_default_enum_with_data_set_orm.php
@@ -14,28 +14,28 @@ namespace App\Factory;
 use Doctrine\ORM\EntityRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\EntityWithEnum;
 use Zenstruck\Foundry\Tests\Fixtures\PHP81\SomeEnum;
 
 /**
  * @extends PersistentProxyObjectFactory<EntityWithEnum>
  *
- * @method        EntityWithEnum|Proxy             create(array|callable $attributes = [])
- * @method static EntityWithEnum|Proxy             createOne(array $attributes = [])
- * @method static EntityWithEnum|Proxy             find(object|array|mixed $criteria)
- * @method static EntityWithEnum|Proxy             findOrCreate(array $attributes)
- * @method static EntityWithEnum|Proxy             first(string $sortedField = 'id')
- * @method static EntityWithEnum|Proxy             last(string $sortedField = 'id')
- * @method static EntityWithEnum|Proxy             random(array $attributes = [])
- * @method static EntityWithEnum|Proxy             randomOrCreate(array $attributes = [])
- * @method static EntityRepository|RepositoryProxy repository()
- * @method static EntityWithEnum[]|Proxy[]         all()
- * @method static EntityWithEnum[]|Proxy[]         createMany(int $number, array|callable $attributes = [])
- * @method static EntityWithEnum[]|Proxy[]         createSequence(iterable|callable $sequence)
- * @method static EntityWithEnum[]|Proxy[]         findBy(array $attributes)
- * @method static EntityWithEnum[]|Proxy[]         randomRange(int $min, int $max, array $attributes = [])
- * @method static EntityWithEnum[]|Proxy[]         randomSet(int $number, array $attributes = [])
+ * @method        EntityWithEnum|Proxy                 create(array|callable $attributes = [])
+ * @method static EntityWithEnum|Proxy                 createOne(array $attributes = [])
+ * @method static EntityWithEnum|Proxy                 find(object|array|mixed $criteria)
+ * @method static EntityWithEnum|Proxy                 findOrCreate(array $attributes)
+ * @method static EntityWithEnum|Proxy                 first(string $sortedField = 'id')
+ * @method static EntityWithEnum|Proxy                 last(string $sortedField = 'id')
+ * @method static EntityWithEnum|Proxy                 random(array $attributes = [])
+ * @method static EntityWithEnum|Proxy                 randomOrCreate(array $attributes = [])
+ * @method static EntityRepository|RepositoryDecorator repository()
+ * @method static EntityWithEnum[]|Proxy[]             all()
+ * @method static EntityWithEnum[]|Proxy[]             createMany(int $number, array|callable $attributes = [])
+ * @method static EntityWithEnum[]|Proxy[]             createSequence(iterable|callable $sequence)
+ * @method static EntityWithEnum[]|Proxy[]             findBy(array $attributes)
+ * @method static EntityWithEnum[]|Proxy[]             randomRange(int $min, int $max, array $attributes = [])
+ * @method static EntityWithEnum[]|Proxy[]             randomSet(int $number, array $attributes = [])
  */
 final class EntityWithEnumFactory extends PersistentProxyObjectFactory
 {
