+++ tests/Fixtures/Maker/expected/can_create_factory_with_embeddable_with_data_set_orm.php
@@ -14,27 +14,27 @@ namespace App\Factory;
 use Doctrine\ORM\EntityRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Contact;
 
 /**
  * @extends PersistentProxyObjectFactory<Contact>
  *
- * @method        Contact|Proxy                    create(array|callable $attributes = [])
- * @method static Contact|Proxy                    createOne(array $attributes = [])
- * @method static Contact|Proxy                    find(object|array|mixed $criteria)
- * @method static Contact|Proxy                    findOrCreate(array $attributes)
- * @method static Contact|Proxy                    first(string $sortedField = 'id')
- * @method static Contact|Proxy                    last(string $sortedField = 'id')
- * @method static Contact|Proxy                    random(array $attributes = [])
- * @method static Contact|Proxy                    randomOrCreate(array $attributes = [])
- * @method static EntityRepository|RepositoryProxy repository()
- * @method static Contact[]|Proxy[]                all()
- * @method static Contact[]|Proxy[]                createMany(int $number, array|callable $attributes = [])
- * @method static Contact[]|Proxy[]                createSequence(iterable|callable $sequence)
- * @method static Contact[]|Proxy[]                findBy(array $attributes)
- * @method static Contact[]|Proxy[]                randomRange(int $min, int $max, array $attributes = [])
- * @method static Contact[]|Proxy[]                randomSet(int $number, array $attributes = [])
+ * @method        Contact|Proxy                        create(array|callable $attributes = [])
+ * @method static Contact|Proxy                        createOne(array $attributes = [])
+ * @method static Contact|Proxy                        find(object|array|mixed $criteria)
+ * @method static Contact|Proxy                        findOrCreate(array $attributes)
+ * @method static Contact|Proxy                        first(string $sortedField = 'id')
+ * @method static Contact|Proxy                        last(string $sortedField = 'id')
+ * @method static Contact|Proxy                        random(array $attributes = [])
+ * @method static Contact|Proxy                        randomOrCreate(array $attributes = [])
+ * @method static EntityRepository|RepositoryDecorator repository()
+ * @method static Contact[]|Proxy[]                    all()
+ * @method static Contact[]|Proxy[]                    createMany(int $number, array|callable $attributes = [])
+ * @method static Contact[]|Proxy[]                    createSequence(iterable|callable $sequence)
+ * @method static Contact[]|Proxy[]                    findBy(array $attributes)
+ * @method static Contact[]|Proxy[]                    randomRange(int $min, int $max, array $attributes = [])
+ * @method static Contact[]|Proxy[]                    randomSet(int $number, array $attributes = [])
  */
 final class ContactFactory extends PersistentProxyObjectFactory
 {
