+++ tests/Fixtures/Maker/expected/can_create_factory_with_static_analysis_annotations_with_data_set_phpstan.php
@@ -14,27 +14,27 @@ namespace App\Factory;
 use Doctrine\ORM\EntityRepository;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
 use Zenstruck\Foundry\Persistence\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\RepositoryDecorator;
 use Zenstruck\Foundry\Tests\Fixtures\Entity\Category;
 
 /**
  * @extends PersistentProxyObjectFactory<Category>
  *
- * @method        Category|Proxy                   create(array|callable $attributes = [])
- * @method static Category|Proxy                   createOne(array $attributes = [])
- * @method static Category|Proxy                   find(object|array|mixed $criteria)
- * @method static Category|Proxy                   findOrCreate(array $attributes)
- * @method static Category|Proxy                   first(string $sortedField = 'id')
- * @method static Category|Proxy                   last(string $sortedField = 'id')
- * @method static Category|Proxy                   random(array $attributes = [])
- * @method static Category|Proxy                   randomOrCreate(array $attributes = [])
- * @method static EntityRepository|RepositoryProxy repository()
- * @method static Category[]|Proxy[]               all()
- * @method static Category[]|Proxy[]               createMany(int $number, array|callable $attributes = [])
- * @method static Category[]|Proxy[]               createSequence(iterable|callable $sequence)
- * @method static Category[]|Proxy[]               findBy(array $attributes)
- * @method static Category[]|Proxy[]               randomRange(int $min, int $max, array $attributes = [])
- * @method static Category[]|Proxy[]               randomSet(int $number, array $attributes = [])
+ * @method        Category|Proxy                       create(array|callable $attributes = [])
+ * @method static Category|Proxy                       createOne(array $attributes = [])
+ * @method static Category|Proxy                       find(object|array|mixed $criteria)
+ * @method static Category|Proxy                       findOrCreate(array $attributes)
+ * @method static Category|Proxy                       first(string $sortedField = 'id')
+ * @method static Category|Proxy                       last(string $sortedField = 'id')
+ * @method static Category|Proxy                       random(array $attributes = [])
+ * @method static Category|Proxy                       randomOrCreate(array $attributes = [])
+ * @method static EntityRepository|RepositoryDecorator repository()
+ * @method static Category[]|Proxy[]                   all()
+ * @method static Category[]|Proxy[]                   createMany(int $number, array|callable $attributes = [])
+ * @method static Category[]|Proxy[]                   createSequence(iterable|callable $sequence)
+ * @method static Category[]|Proxy[]                   findBy(array $attributes)
+ * @method static Category[]|Proxy[]                   randomRange(int $min, int $max, array $attributes = [])
+ * @method static Category[]|Proxy[]                   randomSet(int $number, array $attributes = [])
  *
  * @phpstan-method        Proxy<Category> create(array|callable $attributes = [])
  * @phpstan-method static Proxy<Category> createOne(array $attributes = [])
@@ -44,7 +44,7 @@ use Zenstruck\Foundry\Tests\Fixtures\Entity\Category;
  * @phpstan-method static Proxy<Category> last(string $sortedField = 'id')
  * @phpstan-method static Proxy<Category> random(array $attributes = [])
  * @phpstan-method static Proxy<Category> randomOrCreate(array $attributes = [])
- * @phpstan-method static RepositoryProxy<Category> repository()
+ * @phpstan-method static RepositoryDecorator<Category> repository()
  * @phpstan-method static list<Proxy<Category>> all()
  * @phpstan-method static list<Proxy<Category>> createMany(int $number, array|callable $attributes = [])
  * @phpstan-method static list<Proxy<Category>> createSequence(iterable|callable $sequence)
